+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - post_check:  Added branch name as arg to is_commits_ahead and is_commits_behind calls.
 
 ### Changed
+- post_process:  Added datetime to destination directory for the archiving of the repo.
 - merge_project:  Added additional logging error statements.
 - process_project:  Added additional logging error statements.